Change position cache invalidation to use first and last key
Before, the key positions cache was invalidated when the position of the
middle C changed. When switching from a 25-key keyboard to a 49-key one,
the posision of the middle C wasn't changed, to moving to one of the 24
added keys didn't work.
Now I use the position of the first and last keys, and if any of those
change a cache reload is triggered.